HSEARCH-3619 Move the complexity of retrieving the current session from HibernateSearchContextService to HibernateOrmMapping
That's another to remove some of the cyclic dependencies.
